Stop coercing compare value to int, use float
Former-commit-id: 04d28b60a4b7fcf31cedd0a3eacc61935cb6ce46

+++ stringMLST.py
@@ -331,7 +331,7 @@ def getMaxCount(alleleCount, fileName):
         secondSupport[loc] = {}
         num_max = []
         num_max2 = []
-        compare = int(re.sub("\*$", "", str(max_n[loc])))
+        compare = float(re.sub("\*$", "", str(max_n[loc])))
         for num in alleleCount[loc]:
             if  alleleCount[loc][num] == compare:
                 if "\*" in str(max_n[loc]):
